+++ GIFT/src/edu/asu/selab/steelartt/common/ta/state/scenarioDefinition/ObjectOfInterest.java
@@ -6,7 +6,7 @@ import java.util.List;
 public class ObjectOfInterest extends Entity {
     private String type;
 
-    public ObjectOfInterest(String id, String description, String type, List<Integer> location, List<AreaOfInterest> areasOfInterest) {
+    public ObjectOfInterest(String id, String description, String type, List<Double> location, List<AreaOfInterest> areasOfInterest) {
         super(id, description, location, areasOfInterest);
         this.type = type;
     }
