+++ GIFT/src/edu/asu/selab/steelartt/net/embedded/message/scenarioDefinition/Trainee.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class Trainee extends Entity {
         private String role;
 
-        public Trainee(String id, String description, String role, List<Integer> location, List<AreaOfInterest> areasOfInterest) {
+        public Trainee(String id, String description, String role, List<Double> location, List<AreaOfInterest> areasOfInterest) {
             super(id, description, location, areasOfInterest);
             this.role = role;
         }
